[Feat] 사유 서버데이터 ui에 입히기
