--- PARD.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,32 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "pard_designsystem",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/Club-PARD/PARD_DesignSystem.git",
-      "state" : {
-        "branch" : "main",
-        "revision" : "9bd32fa805af2f92af95f4ea0fdef81945547651"
-      }
-    },
-    {
-      "identity" : "snapkit",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/SnapKit/SnapKit.git",
-      "state" : {
-        "revision" : "2842e6e84e82eb9a8dac0100ca90d9444b0307f4",
-        "version" : "5.7.1"
-      }
-    },
-    {
-      "identity" : "then",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/devxoul/Then.git",
-      "state" : {
-        "revision" : "d41ef523faef0f911369f79c0b96815d9dbb6d7a",
-        "version" : "3.0.0"
-      }
-    }
-  ],
-  "version" : 2
-}
